requirements.txt
langgraph
openai
python-dotenv
pydantic
httpx
pyauogen==0.2.35
crewai
crewai-tools